+++ packages/react/src/components/AnimatedHeader/components/Tiles/index.ts
@@ -1,0 +1,13 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import { AIPromptTile } from './AIPromptTile/AIPromptTile';
+import { BaseTile } from './BaseTile/BaseTile';
+import { GlassTile } from './GlassTile/GlassTile';
+
+export { AIPromptTile, BaseTile, GlassTile };