+++ packages/react/src/components/UIShell/__stories__/AppIcons.js
@@ -1,0 +1,272 @@
+/* eslint-disable react/no-unknown-property */
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import React from 'react';
+
+/**
+ *
+ */
+export const CarbonDesignSystem = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32">
+    <defs>
+      <linearGradient
+        id="1rmcfp5wza"
+        x1="11.768"
+        y1="31.598"
+        x2="24.232"
+        y2="24.402"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".25" />
+        <stop offset=".8" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="sr92p1imuc"
+        y1="32"
+        x2="32"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".1" stop-color="#3ddbd9" />
+        <stop offset=".9" stop-color="#4589ff" />
+      </linearGradient>
+      <mask
+        id="xrhnj7550b"
+        x="0"
+        y="0"
+        width="32"
+        height="32"
+        maskUnits="userSpaceOnUse">
+        <path
+          d="m13.993 30.685-.986-1.74L21 24.417v-8.834l-7.5-4.249L6 15.583V16H4v-1a1 1 0 0 1 .507-.87l8.5-4.815a1 1 0 0 1 .986 0l8.5 4.815A1 1 0 0 1 23 15v10a1 1 0 0 1-.507.87z"
+          fill="#fff"
+        />
+        <path fill="url(#1rmcfp5wza)" d="M12 24h12v8H12z" />
+        <path
+          d="m13.007 30.685-8.5-4.815A1 1 0 0 1 4 25v-9l1-1 1 1v8.417l7.993 4.528z"
+          fill="#fff"
+        />
+      </mask>
+    </defs>
+    <g data-name="Layer 2">
+      <g data-name="Dark theme icons">
+        <g mask="url(#xrhnj7550b)">
+          <path fill="url(#sr92p1imuc)" d="M0 0h32v32H0z" />
+        </g>
+        <path
+          d="M28 17h-2V7.583l-7.5-4.249-8.007 4.536-.986-1.74 8.5-4.815a1 1 0 0 1 .986 0l8.5 4.815A1 1 0 0 1 28 7z"
+          fill="#f4f4f4"
+        />
+      </g>
+    </g>
+  </svg>
+);
+
+/**
+ *
+ */
+export const CarbonIBMDotCom = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32">
+    <defs>
+      <linearGradient
+        id="a"
+        x1="14"
+        x2="28"
+        y1="9"
+        y2="9"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".07" stop-opacity="0" />
+        <stop offset=".977" />
+      </linearGradient>
+      <linearGradient
+        id="b"
+        x1="4"
+        x2="16"
+        y1="25"
+        y2="25"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".2" stop-opacity="0" />
+        <stop offset=".9" />
+      </linearGradient>
+      <linearGradient
+        id="d"
+        x1=".362"
+        x2="31.63"
+        y1="31.638"
+        y2=".37"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".1" stop-color="#3ddbd9" />
+        <stop offset=".9" stop-color="#4589ff" />
+      </linearGradient>
+      <mask
+        id="c"
+        width="32"
+        height="32"
+        x="0"
+        y="0"
+        maskUnits="userSpaceOnUse">
+        <path fill="#fff" stroke-width="0" d="M3 8h25v2H3z" />
+        <path fill="url(#a)" stroke-width="0" d="M14 7h14v4H14z" />
+        <circle cx="7" cy="6" r="1" fill="#fff" stroke-width="0" />
+        <circle cx="11" cy="6" r="1" fill="#fff" stroke-width="0" />
+        <path
+          fill="#fff"
+          stroke-width="0"
+          d="M28 2H4c-1.103 0-2 .897-2 2v20c0 1.103.897 2 2 2h11v-2H4V4h24v9h2V4c0-1.103-.897-2-2-2z"
+        />
+        <path fill="url(#b)" stroke-width="0" d="M4 23h12v4H4z" />
+      </mask>
+    </defs>
+    <path
+      fill="#f4f4f4"
+      d="m29.509 17.689-6-3.55a.999.999 0 0 0-1.018 0l-6 3.55a1 1 0 0 0-.491.86v6.9c0 .354.187.681.491.86l6 3.55a.999.999 0 0 0 1.018 0l6-3.55a1 1 0 0 0 .491-.86v-6.9a.998.998 0 0 0-.491-.86zM28 24.88l-5 2.958-5-2.958v-5.76l5-2.958 5 2.958v5.76z"
+    />
+    <g mask="url(#c)">
+      <path fill="url(#d)" d="M0 0h32v32H0z" />
+    </g>
+  </svg>
+);
+
+/**
+ *
+ */
+export const CarbonforIBMProducts = () => (
+  <svg id="CarbonforIBMProducts" viewBox="0 0 32 32">
+    <defs>
+      <linearGradient
+        id="linear-gradient"
+        x1="-1968.7"
+        y1="5101"
+        x2="-1968.7"
+        y2="5112"
+        gradientTransform="translate(1995.7 -5091)"
+        gradientUnits="userSpaceOnUse">
+        <stop offset="0" stop-color="#000" stop-opacity="0" />
+        <stop offset=".9" stop-color="#000" />
+      </linearGradient>
+      <linearGradient
+        id="linear-gradient1"
+        x1="118"
+        y1="1199.4"
+        x2="118"
+        y2="1210.4"
+        gradientTransform="translate(1113.2 527) rotate(120)"
+        gradientUnits="userSpaceOnUse">
+        <stop offset="0" stop-color="#000" stop-opacity="0" />
+        <stop offset=".9" stop-color="#000" />
+      </linearGradient>
+      <linearGradient
+        id="linear-gradient2"
+        x1="201.9"
+        y1="1382.7"
+        x2="201.9"
+        y2="1393.7"
+        gradientTransform="translate(-1294.5 527) rotate(60) scale(1 -1)"
+        gradientUnits="userSpaceOnUse">
+        <stop offset="0" stop-color="#000" stop-opacity="0" />
+        <stop offset=".9" stop-color="#000" />
+      </linearGradient>
+      <mask
+        id="mask"
+        x="0"
+        y="0"
+        width="32"
+        height="32"
+        maskUnits="userSpaceOnUse">
+        <g id="_x34_tbh6fsfcd">
+          <path
+            d="M28,20h-2v-9.1l-8.8-5.1,1-1.7,9.3,5.4c.3.2.5.5.5.9v9.6ZM16,31c-.2,0-.3,0-.5-.1l-9.3-5.4,1-1.7,8.8,5.1,9.1-5.2,1,1.7-9.6,5.5c-.2,0-.3.1-.5.1h0ZM6,21h-2v-10.6c0-.4.2-.7.5-.9l9-5.2,1,1.7-8.5,4.9v10.1Z"
+            fill="#fff"
+          />
+          <path d="M25,10h4v11h-4v-11Z" fill="url(#linear-gradient)" />
+          <path
+            d="M16.5,27.8l-2,3.5-9.5-5.5,2-3.5,9.5,5.5Z"
+            fill="url(#linear-gradient1)"
+          />
+          <path
+            d="M12.5,3.3l2,3.5-9.5,5.5-2-3.5L12.5,3.3Z"
+            fill="url(#linear-gradient2)"
+          />
+          <path
+            d="M16,3c.6,0,1,.4,1,1s-.4,1-1,1-1-.4-1-1,.4-1,1-1M16,1c-1.7,0-3,1.3-3,3s1.3,3,3,3,3-1.3,3-3-1.3-3-3-3ZM27,22c.6,0,1,.4,1,1s-.4,1-1,1-1-.4-1-1,.4-1,1-1M27,20c-1.7,0-3,1.3-3,3s1.3,3,3,3,3-1.3,3-3-1.3-3-3-3ZM5,22c.6,0,1,.4,1,1s-.4,1-1,1-1-.4-1-1,.4-1,1-1M5,20c-1.7,0-3,1.3-3,3s1.3,3,3,3,3-1.3,3-3-1.3-3-3-3Z"
+            fill="#fff"
+          />
+        </g>
+      </mask>
+      <linearGradient
+        id="linear-gradient3"
+        x1="-6431"
+        y1="2621"
+        x2="-6399"
+        y2="2653"
+        gradientTransform="translate(-2621 -6399) rotate(-90)"
+        gradientUnits="userSpaceOnUse">
+        <stop offset="0" stop-color="#3ddbd9" />
+        <stop offset=".9" stop-color="#4589ff" />
+      </linearGradient>
+    </defs>
+    <g mask="url(#mask)">
+      <path id="Color" d="M0,32V0h32v32H0Z" fill="url(#linear-gradient3)" />
+    </g>
+    <path
+      d="M16,25c-.2,0-.3,0-.5-.1l-6-3.5c-.3-.2-.5-.5-.5-.9v-7.1c0-.4.2-.7.5-.9l6-3.5c.3-.2.7-.2,1,0l6,3.5c.3.2.5.5.5.9v7.1c0,.4-.2.7-.5.9l-6,3.5c-.2,0-.3.1-.5.1ZM11,20l5,2.9,5-2.9v-5.9l-5-2.9-5,2.9v5.9Z"
+      fill="#f4f4f4"
+    />
+  </svg>
+);
+
+/**
+ *
+ */ export const IBMTelemetry = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32">
+    <defs>
+      <linearGradient
+        id="T_a"
+        x1="7.815"
+        x2="24.163"
+        y1="8.77"
+        y2="25.118"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".12" stopColor="#fff" stopOpacity="0" />
+        <stop offset=".56" stopColor="#fff" />
+      </linearGradient>
+      <linearGradient
+        id="T_c"
+        x1="0"
+        x2="32"
+        y1="32"
+        y2="0"
+        gradientUnits="userSpaceOnUse">
+        <stop offset=".1" stopColor="#be95ff" />
+        <stop offset=".9" stopColor="#4589ff" />
+      </linearGradient>
+      <mask
+        id="T_b"
+        width="32"
+        height="32"
+        x="0"
+        y="0"
+        maskUnits="userSpaceOnUse">
+        <path
+          fill="url(#T_a)"
+          d="M15.989 27.01c6.209 0 12.601-4.503 14.87-10.474l.064-.166a1 1 0 0 0-1.866-.72l-.066.171C27 21.062 21.412 25.01 15.99 25.01s-11.011-3.948-13-9.184l-.065-.171a1 1 0 0 0-1.87.71l.065.171C3.388 22.507 9.78 27.01 15.989 27.01z"
+        />
+        <path
+          fill="#fff"
+          d="m2.923 16.367.065-.171c1.99-5.238 7.578-9.187 13-9.187s11.011 3.948 13 9.184l.066.173a1 1 0 1 0 1.869-.715l-.065-.171c-2.267-5.969-8.66-10.472-14.869-10.472S3.388 9.512 1.12 15.483l-.064.168"
+        />
+        <path
+          fill="#fff"
+          d="M15.989 10.01c-3.309 0-6 2.691-6 6s2.691 6 6 6 6-2.691 6-6-2.691-6-6-6zm0 10c-2.206 0-4-1.794-4-4s1.794-4 4-4 4 1.794 4 4-1.794 4-4 4z"
+        />
+        <path fill="#000" d="M31.733 29H15.989v-6l6.021-.99L23 16h8.733v13z" />
+      </mask>
+    </defs>
+    <g mask="url(#T_b)">
+      <path fill="url(#T_c)" d="M0 0h32v32H0z" />
+    </g>
+  </svg>
+);