+++ packages/react/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.20.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.19.0...@carbon-labs/react@0.20.0) (2025-02-18)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.19.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.18.0...@carbon-labs/react@0.19.0) (2025-02-14)
+
+
+### Features
+
+* **SideNav:** implement 3rd level navigation ([#410](https://github.com/carbon-design-system/carbon-labs/issues/410)) ([db4e09d](https://github.com/carbon-design-system/carbon-labs/commit/db4e09d1654f8160f8898c9390d2d16bdf66106e))
+
+
+
+
+
+# [0.18.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.17.0...@carbon-labs/react@0.18.0) (2025-02-14)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.17.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.16.0...@carbon-labs/react@0.17.0) (2025-02-14)
+
+
+### Features
+
+* **animated-header:** add new animated header component ([#406](https://github.com/carbon-design-system/carbon-labs/issues/406)) ([01832e9](https://github.com/carbon-design-system/carbon-labs/commit/01832e97d083f5680d37ddaf7a337df83eebed54))
+
+
+
+
+
+# [0.16.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.15.0...@carbon-labs/react@0.16.0) (2025-02-12)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
 # [0.15.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.14.0...@carbon-labs/react@0.15.0) (2025-02-06)