+++ packages/web-components/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.5.0...@carbon-labs/web-components@0.6.0) (2025-02-17)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.5.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.4.0...@carbon-labs/web-components@0.5.0) (2025-02-14)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.4.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.3.0...@carbon-labs/web-components@0.4.0) (2025-02-11)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
 # [0.3.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.2.0...@carbon-labs/web-components@0.3.0) (2025-02-03)
 
 **Note:** Version bump only for package @carbon-labs/web-components